refactor - add db_protocol.py
