+++ database/feature_points_position_db.py
@@ -1,30 +1,17 @@
 import MySQLdb
+from database.db_protocol import DBProtocol
 
-class FeaturePointsPositionDB:
+class FeaturePointsPositionDB(DBProtocol):
     def __init__(self):
-        self.connection = MySQLdb.connect(
-            host='db',
-            port=3306,
-            user='docker',
-            passwd='docker',
-            db='feature_points_position'
-            )
-        self.connection.autocommit(False)
-        self.cursor = self.connection.cursor()
+        db_name = 'feature_points_position'
+        create_db_sql = "CREATE TABLE IF NOT EXISTS feature_points_position(id INT, x FLOAT, y FLOAT, z FLOAT)"
 
-        sql = "CREATE TABLE IF NOT EXISTS feature_points_position(id INT, x FLOAT, y FLOAT, z FLOAT)"
-        self.cursor.execute(sql)
+        super().__init__(db_name, create_db_sql)
 
-    def close(self):
-        self.connection.commit()
-        self.connection.close()
+    def create(self, fp_id, x, y, z):
+        sql = "INSERT INTO feature_points_position(id, x, y, z) VALUES ({}, {}, {}, {})".format(fp_id, x, y, z)
+        super().create(sql)
 
-    def create(self, id, x, y, z):
-        sql = "INSERT INTO feature_points_position(id, x, y, z) VALUES ({}, {}, {}, {})".format(id, x, y, z)
-        self.cursor.execute(sql)
-        self.connection.commit()
-
-    def find(self, id):
-        sql = "SELECT * FROM feature_points_position WHERE id={}".format(id)
-        self.cursor.execute(sql)
-        return self.cursor.fetchall()[0]
+    def find(self, fp_id):
+        sql = "SELECT * FROM feature_points_position WHERE id={}".format(fp_id)
+        return super().find(sql)
